+++ kanleitos/src/main/java/br/com/kanleitos/repository/AlaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.kanleitos.models.Ala;
 
 @Repository
-public interface AlaRepository extends CrudRepository<Ala, Integer> {
+public interface AlaRepository extends CrudRepository<Ala, Long> {
 
 }
