+++ kanleitos/src/main/java/br/com/kanleitos/repository/EnfermariaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.kanleitos.models.Enfermaria;
 
 @Repository
-public interface EnfermariaRepository extends CrudRepository<Enfermaria, Integer> {
+public interface EnfermariaRepository extends CrudRepository<Enfermaria, Long> {
 
 }
