+++ kanleitos/src/main/java/br/com/kanleitos/repository/UsuarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.kanleitos.models.Usuario;
 
 @Repository
-public interface UsuarioRepository extends CrudRepository<Usuario, Integer> {
+public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
 	Usuario findByLogin (String login);
 }
